Fix view logs admin screen
