+++ app/logger/server/index.js
@@ -1,7 +1,2 @@
-import { Logger } from '../../../server/lib/logger/Logger';
-import './streamer';
-
 // TODO there are imports pointing to this file still, ideally we should point everything to "/server/lib/logger/Logger" and remove this file
-export {
-	Logger,
-};
+export { Logger } from '../../../server/lib/logger/Logger';
